% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS_API_Key.R
\name{MARS_API_Key}
\alias{MARS_API_Key}
\title{MARS_API_Key}
\usage{
MARS_API_Key(key)
}
\arguments{
\item{key}{your API Key obtained from (https://mymarketnews.ams.usda.gov/mymarketnews-api)}
}
\description{
This function sets a variable in the R environment for your MARS API Key
}
\examples{
#set your API Key
MARS_API_Key(key = "Your API Key")

}
