% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS_Table_Directory.R
\name{MARS_Table_Directory}
\alias{MARS_Table_Directory}
\title{MARS_Table_Directory}
\usage{
MARS_Table_Directory()
}
\arguments{
\item{takes}{no arguments}
}
\description{
This function pulls a dataframe for all of the reports, their corresponding slug ids , and relevant metadata.
}
\examples{
#Pull the report/table directory

MARS_Table_Directory()

}
