% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS_Table_Pull.R
\name{MARS_Table_Pull}
\alias{MARS_Table_Pull}
\title{MARS_Table_Pull}
\usage{
MARS_Table_Pull(slugid)
}
\arguments{
\item{slugid}{the slug id for the report containing the data you want to pull}
}
\description{
This function pulls a dataframe for a single report/slug id based on user input.
}
\examples{
#Pull the report/table directory
Df <- MARS_Table_Pull(slugid)


}
