% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS_Table_Pull.R
\name{MARS_Table_Pull}
\alias{MARS_Table_Pull}
\title{MARS_Table_Pull}
\usage{
MARS_Table_Pull(slugid, section, date)
}
\arguments{
\item{slugid}{The slug id for the report containing the data you want to pull}

\item{section}{Optional, the report section}

\item{date}{Optional, the report date or range of dates}
}
\description{
This function pulls a dataframe for a single report/slug id.
}
\examples{
#Pull the report/table directory
Df <- MARS_Table_Pull(slugid, section, date)


}
